chore: test rendering image popup from component
Try to render the image in a popup from the LocationImage component. It
looks like renderToStaticMarkup should be used, except that has been
replaced with renderToPipeableStream. However, an error happens when
importing renderToPipeableStream. In general, this error and other
errors have been reported within the past week. It seems like this is
because of Next.js upgrading to version 14. I think it would be good to
revisit this later, after breaking changes will of had time to be fixed.