feat: create location guesser game (#2)
* feat: display region map and location image

Make it so that the game displays a region map and location image.

* refactor: organize components folder structure

Move components around and rename them in the components folder.

* refactor: split Map component into Map and GameMap

Extract the basic Map part of the Map component, and then have the new
GameMap component be for the actual game. This is so that a map can be
displayed on other parts of the website, without the game functionality.

* refactor: move styling around

Move components and css styling around.

* refactor: make LocationImage apply className

Make the LocationImage apply the className styling now instead of
GameLocationImage. This is done by having GameLocationImage pass a prop
of className to LocationImage.

* feat: resize game image div with useEffect

Resize the game image div outside of CSS.

* feat: square game image depends on width or height

The game location image can be made to be square, but depends only on
either width or height.

* feat: square game location image works

The game location image is now correctly a square, and the size depends
on the screen size.

* refactor: remove comments of unused code

Remove comments of old unused code.

* feat: add and remove guess markers

After a guess is made, display the location of the guess and the correct
location. A line is added between the two markers. Pressing a "Next"
button removes the two markers and the line, and allows the player to
guess again.

* feat: display correct location image popup

Display a popup of the correct location after guessing, above the
correct location marker.

* feat: cycle through locations

After a location is guessed, move to the next location to guess.

* refactor: load in game data in the game container

Make the game container load in the game data, so that each component
doesn't have to load in the game data (the game container shares the
loaded game data with its components).

* feat: randomly choose area locations

Randomly choose a new location which hasn't been chosen yet.

* chore: test rendering image popup from component

Try to render the image in a popup from the LocationImage component. It
looks like renderToStaticMarkup should be used, except that has been
replaced with renderToPipeableStream. However, an error happens when
importing renderToPipeableStream. In general, this error and other
errors have been reported within the past week. It seems like this is
because of Next.js upgrading to version 14. I think it would be good to
revisit this later, after breaking changes will of had time to be fixed.

* refactor: add util function to load map data

Add a function in a new utils folder, that loads map data.

* feat: display a map on the home screen

Display a map using the Map component on the home screen.

* feat: display home page title in middle of screen

Display the title of the home page in the middle of the screen.

* feat: make location images use natural dimensions

Location images now use their natural dimensions when rendering.

* feat: make location image use natural dimensions

The location image will now keep its natural dimensions, and scale to be
a good size (max width and max height currently 75% of the game
container).

* feat: add border radius to popup image

Clean up the popup image code, and give the image of a popup a border
radius.

* feat: add a green theme to the home page

Make the home page have a green theme. Text is light green, and the
background color is a darker green.

* feat: add navbar link components

Add components for the Navbar links.

* feat: add a text container component

Add a container component that is good for displaying text. It is being
used so far for the About page.

* refactor: use container component for home page

Use the container component good for text for the home page text.

* feat: change map zoom level based on screen width

Make it so that if the screen width is small, the initial map zoom is
further out. This is done mainly for mobile devices.

* feat: add game results and guess distances

When the game finishes, display game results. A summary of guess
distances is shown (total, closest, and furthest). Also show the guess
distance after each guess.

* feat: add more locations

Add more locations that can be guessed.