feat: add submit button for map location guess (#37)
Add a submit button for guessing a location. This is so that the guess
marker position can be corrected. Also, it prevents accidently
submitting on a device such as mobile when trying to zoom in.

As a side effect of this fix, the correct marker popup now uses the
LocationImage component instead of a hardcoded <img> content, which is a
good change.