feat: avoid repeating locations (#59)
* feat: avoid repeating locations

Make it so that a certain number of games have to pass before a location
can appear again.

* fix: remove max image file size check change path

This has to be done since the workflow is set to be required. If a
change is made in a directory not under the path specified for the
workflow, then the workflow never triggers and the "required" check
requirement is never fulfilled. A "required" check means it must run and
pass. So, if the check never runs, even though if it would run it would
pass, the pull request still can't be submitted.

* refactor: simplify max image file size check code

Simplify the check "on" part to be just "on: pull_request".